resolve implicit type conversion warnings

+++ airport.c
@@ -195,7 +195,7 @@ int airport_cmd_nearest(CMDPARAMS)
 		if (airports[i].enabled) {
 			dx = airports[i].pos.x - playerx;
 			dy = airports[i].pos.y - playery;
-			aps[num].distance = sqrt(dx * dx + dy * dy);
+			aps[num].distance = sqrtf(dx * dx + dy * dy);
 			aps[num].index = i;
 			num++;
 		}
